Move HasLocallyFree trait to interpreter where it's used
