+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/storage/implicits.scala
@@ -4,6 +4,7 @@ import cats.implicits._
 import coop.rchain.models.Channel.ChannelInstance.Quote
 import coop.rchain.models._
 import coop.rchain.models.implicits.mkProtobufInstance
+import coop.rchain.rholang.interpreter.HasLocallyFree
 import coop.rchain.rholang.interpreter.SpatialMatcher._
 import coop.rchain.rholang.interpreter.implicits._
 import coop.rchain.rspace.{Serialize, Match => StorageMatch}
